# api/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Pulse Pager API
  version: 1.0.0
  description: >
    Pulse Pager public REST API (v1). This spec is the single source of truth (RFC-012):
    the Go server types/stubs and the TypeScript client types are generated from it.
    Auth-plane redirects (/auth/{provider}/login, /auth/refresh, /auth/logout) are
    not modelled here; this is the JSON resource API under /api/v1.
servers:
  - url: /api/v1
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /orgs/{orgId}/monitors/{id}/check:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/Id"
    post:
      operationId: checkNow
      tags: [monitors]
      summary: >
        Trigger a check now. It enqueues one probe per region (like a scheduled
        check) and returns 202 immediately; the per-region progress shows up in the
        monitor's live region states (poll getMonitorRegionStates). 409 if a check is
        already running, the monitor is disabled, or you are within the per-monitor
        cooldown; 429 if you are over the sustained manual-check budget.
      responses:
        "202":
          description: accepted; the check was enqueued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckNowAccepted" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pulse_at
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema: { type: string }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: not signed in
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    TooManyRequests:
      description: rate limited; retry after the indicated delay
      headers:
        Retry-After:
          schema: { type: integer }
          description: seconds to wait before retrying
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Forbidden:
      description: not allowed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    FailureReason:
      type: string
      enum: [connection_error, timeout, status_mismatch, latency_exceeded, body_assertion_failed, blocked_target, cert_expired, cert_expiring_soon, cert_invalid]
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/Error" }
...
    RegionState:
      type: object
      required: [region, state, updated_at]
      properties:
        region: { type: string }
        state: { type: string, enum: [scheduled, running, done, failed] }
        healthy: { type: boolean, nullable: true }
        status_code: { type: integer, nullable: true }
        latency_ms: { type: integer, nullable: true }
        failure_reason:
          allOf: [{ $ref: "#/components/schemas/FailureReason" }]
          nullable: true
        updated_at: { type: string, format: date-time }
...
    CheckNowAccepted:
      type: object
      required: [monitor_id, regions]
      properties:
        monitor_id: { type: string }
        regions:
          type: array
          items: { $ref: "#/components/schemas/RegionState" }
...